pyproject.toml
[tool.poetry]
name = "bro"
version = "0.1.0"
description = "Financial bro assistant"
authors = ["Miguel Cárdenas <miguelecsx@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.11"
pydantic = "^2.11.4"
pandas = "^2.2.3"
pydantic-settings = "^2.9.1"
seaborn = "^0.13.2"
scikit-learn = "^1.6.1"
statsmodels = "^0.14.4"
mcp = {extras = ["cli"], version = "^1.7.1"}
starlette = "^0.46.2"
uvicorn = "^0.34.2"
pillow = "^11.2.1"
hmmlearn = "^0.3.3"
tqdm = "^4.67.1"
yfinance = "^0.2.58"
torch = "^2.7.0"
prophet = "^1.1.6"
plotly = "^6.0.1"
langchain-mcp-adapters = "^0.0.10"
langchain-google-genai = "^2.1.4"
langgraph = "^0.4.1"
langchain = "^0.3.25"
fastapi = "^0.115.12"
arch = "^7.2.0"
xgboost = "^3.0.0"
pmdarima = "^2.0.4"
lightgbm = "^4.6.0"
catboost = "^1.2.8"
pykalman = {version = "^0.10.1", python = ">=3.11,<3.14"}

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
